storage: move MVCC key/encoding logic into separate file
This splits MVCC keys and their encoding into a separate file, following
the same convention as `engine_key.go`. There is additional decoding
logic in `enginepb/decode.go` which has not been moved -- this was
originally to avoid a RocksDB CGo dependency for these functions, but
Pebble still has a (tiny) CGo dependency so we leave them for now.

Release note: None
