+++ pkg/storage/BUILD.bazel
@@ -19,6 +19,7 @@ go_library(
         "multi_iterator.go",
         "mvcc.go",
         "mvcc_incremental_iterator.go",
+        "mvcc_key.go",
         "mvcc_logical_ops.go",
         "open.go",
         "pebble.go",
@@ -105,6 +106,7 @@ go_test(
         "multi_iterator_test.go",
         "mvcc_history_test.go",
         "mvcc_incremental_iterator_test.go",
+        "mvcc_key_test.go",
         "mvcc_logical_ops_test.go",
         "mvcc_stats_test.go",
         "mvcc_test.go",
