+++ pkg/storage/mvcc.go
@@ -57,14 +57,6 @@ const (
 	maxIntentsPerWriteIntentErrorDefault = 5000
 )
 
-var (
-	// MVCCKeyMax is a maximum mvcc-encoded key value which sorts after
-	// all other keys.
-	MVCCKeyMax = MakeMVCCMetadataKey(roachpb.KeyMax)
-	// NilKey is the nil MVCCKey.
-	NilKey = MVCCKey{}
-)
-
 var minWALSyncInterval = settings.RegisterDurationSetting(
 	settings.TenantWritable,
 	"rocksdb.min_wal_sync_interval",
@@ -99,107 +91,6 @@ func MakeValue(meta enginepb.MVCCMetadata) roachpb.Value {
 
 func emptyKeyError() error {
 	return errors.Errorf("attempted access to empty key")
-}
-
-// MVCCKey is a versioned key, distinguished from roachpb.Key with the addition
-// of a timestamp.
-type MVCCKey struct {
-	Key       roachpb.Key
-	Timestamp hlc.Timestamp
-}
-
-// MakeMVCCMetadataKey creates an MVCCKey from a roachpb.Key.
-func MakeMVCCMetadataKey(key roachpb.Key) MVCCKey {
-	return MVCCKey{Key: key}
-}
-
-// Next returns the next key.
-func (k MVCCKey) Next() MVCCKey {
-	ts := k.Timestamp.Prev()
-	if ts.IsEmpty() {
-		return MVCCKey{
-			Key: k.Key.Next(),
-		}
-	}
-	return MVCCKey{
-		Key:       k.Key,
-		Timestamp: ts,
-	}
-}
-
-// Less compares two keys.
-func (k MVCCKey) Less(l MVCCKey) bool {
-	if c := k.Key.Compare(l.Key); c != 0 {
-		return c < 0
-	}
-	if !k.IsValue() {
-		return l.IsValue()
-	} else if !l.IsValue() {
-		return false
-	}
-	return l.Timestamp.Less(k.Timestamp)
-}
-
-// Equal returns whether two keys are identical.
-func (k MVCCKey) Equal(l MVCCKey) bool {
-	return k.Key.Compare(l.Key) == 0 && k.Timestamp.EqOrdering(l.Timestamp)
-}
-
-// IsValue returns true iff the timestamp is non-zero.
-func (k MVCCKey) IsValue() bool {
-	return !k.Timestamp.IsEmpty()
-}
-
-// EncodedSize returns the size of the MVCCKey when encoded.
-func (k MVCCKey) EncodedSize() int {
-	n := len(k.Key) + 1
-	if k.IsValue() {
-		// Note that this isn't quite accurate: timestamps consume between 8-13
-		// bytes. Fixing this only adjusts the accounting for timestamps, not the
-		// actual on disk storage.
-		n += int(MVCCVersionTimestampSize)
-	}
-	return n
-}
-
-// String returns a string-formatted version of the key.
-func (k MVCCKey) String() string {
-	if !k.IsValue() {
-		return k.Key.String()
-	}
-	return fmt.Sprintf("%s/%s", k.Key, k.Timestamp)
-}
-
-// Format implements the fmt.Formatter interface.
-func (k MVCCKey) Format(f fmt.State, c rune) {
-	fmt.Fprintf(f, "%s/%s", k.Key, k.Timestamp)
-}
-
-// Len returns the size of the MVCCKey when encoded. Implements the
-// pebble.Encodeable interface.
-//
-// TODO(itsbilal): Reconcile this with EncodedSize. Would require updating MVCC
-// stats tests to reflect the more accurate lengths provided by this function.
-func (k MVCCKey) Len() int {
-	const (
-		timestampSentinelLen      = 1
-		walltimeEncodedLen        = 8
-		logicalEncodedLen         = 4
-		syntheticEncodedLen       = 1
-		timestampEncodedLengthLen = 1
-	)
-
-	n := len(k.Key) + timestampEncodedLengthLen
-	if !k.Timestamp.IsEmpty() {
-		n += timestampSentinelLen + walltimeEncodedLen
-		if k.Timestamp.Logical != 0 || k.Timestamp.Synthetic {
-			n += logicalEncodedLen
-		}
-		if k.Timestamp.Synthetic {
-			n += syntheticEncodedLen
-		}
-	}
-	return n
 }
 
 // MVCCKeyValue contains the raw bytes of the value for a key.
