+++ pkg/storage/mvcc_key_test.go
@@ -1,0 +1,60 @@
+// Copyright 2022 The Cockroach Authors.
+//
+// Use of this software is governed by the Business Source License
+// included in the file licenses/BSL.txt.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0, included in the file
+// licenses/APL.txt.
+
+package storage
+
+import (
+	"math"
+	"reflect"
+	"sort"
+	"testing"
+
+	"github.com/cockroachdb/cockroach/pkg/roachpb"
+	"github.com/cockroachdb/cockroach/pkg/util/hlc"
+	"github.com/cockroachdb/cockroach/pkg/util/leaktest"
+	"github.com/cockroachdb/cockroach/pkg/util/log"
+	"github.com/cockroachdb/cockroach/pkg/util/shuffle"
+)
+
+// Verify the sort ordering of successive keys with metadata and
+// versioned values. In particular, the following sequence of keys /
+// versions:
+//
+// a
+// a<t=max>
+// a<t=1>
+// a<t=0>
+// a\x00
+// a\x00<t=max>
+// a\x00<t=1>
+// a\x00<t=0>
+func TestMVCCKeys(t *testing.T) {
+	defer leaktest.AfterTest(t)()
+	defer log.Scope(t).Close(t)
+	aKey := roachpb.Key("a")
+	a0Key := roachpb.Key("a\x00")
+	keys := mvccKeys{
+		mvccKey(aKey),
+		mvccVersionKey(aKey, hlc.Timestamp{WallTime: math.MaxInt64}),
+		mvccVersionKey(aKey, hlc.Timestamp{WallTime: 1}),
+		mvccVersionKey(aKey, hlc.Timestamp{Logical: 1}),
+		mvccKey(a0Key),
+		mvccVersionKey(a0Key, hlc.Timestamp{WallTime: math.MaxInt64}),
+		mvccVersionKey(a0Key, hlc.Timestamp{WallTime: 1}),
+		mvccVersionKey(a0Key, hlc.Timestamp{Logical: 1}),
+	}
+	sortKeys := make(mvccKeys, len(keys))
+	copy(sortKeys, keys)
+	shuffle.Shuffle(sortKeys)
+	sort.Sort(sortKeys)
+	if !reflect.DeepEqual(sortKeys, keys) {
+		t.Errorf("expected keys to sort in order %s, but got %s", keys, sortKeys)
+	}
+}
